✨ feat(Project): date_selection
The system will display a menu as below for the user to choose from.   	1.Date (today) 	2.Date (tomorrow) 	3.Date (the day after tomorrow) 	4.Exit  The user can input 1-3 to choose a date if he/she wants to book classroom on that day, or 4 to quit the booking system. Here Date in the menu should be replaced by real day. For example, if today is 20 December 2024, you should put 20/12/2024 or 2024.12.20 there.